src/requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
asyncio
aiohttp==3.9.1
pydantic==2.5.2
rich==13.7.0
loguru==0.7.2 